+++ src/main/java/io/github/kings1990/plugin/fastrequest/configurable/CommonConfigurable.java
@@ -48,7 +48,7 @@ public class CommonConfigurable extends AbstractConfigConfigurable {
 
     @Override
     public String getDisplayName() {
-        return "Restful Fast Request";
+        return "Fast Request";
     }
 
     @Override